% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSARP.r
\docType{package}
\name{rSARP}
\alias{rSARP}
\alias{rSARP-package}
\title{Create, modify, critique, and report against search and rescue plans using \pkg{rSARP}.}
\description{
rSARP allows you to create and manage search and rescue search grid plans using CSV
files and the graphic output from R.  The package includes 3 main functions.  The
\code{searchme()} function provides modeling by sector and is the core of the package.
The \code{searchstatus()} function provides graphic reporting against the core plan and is
typically used several times within a planning cycle as plans are modified. The
\code{bestsearch()} function is used a the beginning of a planning cycle to
determine the best search methods to use for a limited number of searchers. The \code{tracking()}
function is used to track task assignments (525s) as they are assigned and completed.
}
\details{
The first function you're likely to need from \pkg{rSARP} is
\code{\link{searchme}}. Otherwise refer to the package documentation to see
how to proceed.
}

