% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Terrains.csv}
\alias{Terrains.csv}
\title{Terrains.csv - typical Data input file for \pkg{rSARP}}
\format{A data frame with 7 rows and 4 variables}
\source{
John Hutcheson
}
\description{
A dataset containing the terrains and key attributes of each
that are relevant to search. The variables are as follows:
}
\details{
\itemize{
  \item Terrain. Name of the Terrain described.  Usually string descriptions that are easy to remember.
  \item Hours. Relative man hours of search required for the terrain. 
  \item AMDR. Average maximum detection range that is typical for the Terrain. 
  \item Speed.  Expected travel speed through the terrain in miles/hr.
}
}

