% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSARP.r
\name{Tracking.csv}
\alias{Tracking.csv}
\title{Tracking.csv - typical Data input file for \code{tracking()}}
\format{A data frame with 33 rows and 5 variables}
\source{
John Hutcheson
}
\description{
A dataset containing the names or assignment numbers of the 525 tasks assigned
within a search, the type of task, the number of people assigned, and the %
completion of the task.  This file is stored in the working directory of R and
must be named Tracking.csv. The variables described in the file are as follows:
}
\details{
\itemize{
  \item Task. Name or assignment number of the task.  Usually Alpha characters excluding I and O if search sector names are used.
  \item PerCentComplete. Relative task completion amount expressed as a percentage. An entry of 25 indicates the task is one quarter complete.
  \item TSearchers. Number of persons assigned to this task.
  \item Type. String representing the category of task this assignment belongs to.  Generally use Initial, Loose, and Tight types, but the user is free to create sub categories as needed.
  \item Comment. Time of Day.  Keyword describing time of search and currently limited to Day or Night.
}
}

