% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Weathers.csv}
\alias{Weathers.csv}
\title{Weathers.csv - typical Data input file for \pkg{rSARP}}
\format{A data frame with 6 rows and 2 variables}
\source{
John Hutcheson
}
\description{
A dataset containing the Weather conditions that affect search and the speed impacts.
The variables are as follows:
}
\details{
\itemize{
  \item Weather. Name of the Weather described.  Usually string descriptions that are easy to remember and descriptive of conditions.
  \item Speed.  Expected travel speed through the Weather in miles/hr.
}
}

